+++ lib/open_pgp/encode/packet_impl.ex
@@ -2,7 +2,7 @@ defimpl OpenPGP.Encode, for: OpenPGP.Packet do
   alias OpenPGP.Packet
   alias OpenPGP.Packet.BodyChunk
 
-  def tag(_), do: raise(".tag/1 not supported by design for #{inspect(@for)}.")
+  def tag(_), do: raise(".tag/1 of protocol #{inspect(@protocol)} not supported by design for #{inspect(@for)}.")
 
   @doc """
   Encode a Packet.