+++ test/open_pgp/integrity_protected_data_packet_test.exs
@@ -3,14 +3,45 @@ defmodule OpenPGP.IntegrityProtectedDataPacketTest do
   doctest OpenPGP.IntegrityProtectedDataPacket
   doctest OpenPGP.Encode.impl_for!(%OpenPGP.IntegrityProtectedDataPacket{})
 
+  alias OpenPGP.Encrypt
   alias OpenPGP.IntegrityProtectedDataPacket, as: IPDPacket
   alias OpenPGP.PublicKeyEncryptedSessionKeyPacket, as: PKESK
 
-  describe ".encrypt/2,3" do
+  describe ".decrypt/2" do
+    @expected_error "Failed to parse Modification Detection Code Packet."
+    @algo {7, "AES with 128-bit key [AES]"}
+    test "raise error if MDC is missing when `use_mdc: true`" do
+      sym_key = :crypto.strong_rand_bytes(16)
+
+      assert %IPDPacket{ciphertext: ciphertext} =
+               Encrypt.encrypt(%IPDPacket{plaintext: "Hello!"},
+                 session_key: sym_key,
+                 session_key_algo: @algo,
+                 use_mdc: false
+               )
+
+      assert_raise RuntimeError, @expected_error, fn ->
+        %IPDPacket{plaintext: "Hello!"} =
+          IPDPacket.decrypt(
+            %IPDPacket{ciphertext: ciphertext},
+            %PKESK{session_key_algo: @algo, session_key_material: {sym_key}},
+            use_mdc: true
+          )
+      end
+    end
+  end
+
+  describe "OpenPGP.Encrypt.encrypt/2" do
     @algo {7, "AES with 128-bit key [AES]"}
     test "encrypt plaintext with AES-128" do
       sym_key = :crypto.strong_rand_bytes(16)
-      assert ciphertext = IPDPacket.encrypt("Hello!", sym_key, @algo, use_mdc: true)
+
+      assert %IPDPacket{ciphertext: ciphertext} =
+               Encrypt.encrypt(%IPDPacket{plaintext: "Hello!"},
+                 session_key: sym_key,
+                 session_key_algo: @algo,
+                 use_mdc: true
+               )
 
       assert %IPDPacket{plaintext: "Hello!"} =
                IPDPacket.decrypt(
@@ -23,7 +54,13 @@ defmodule OpenPGP.IntegrityProtectedDataPacketTest do
     @algo {8, "AES with 192-bit key"}
     test "encrypt plaintext with AES-192" do
       sym_key = :crypto.strong_rand_bytes(24)
-      assert ciphertext = IPDPacket.encrypt("Hello!", sym_key, @algo, use_mdc: true)
+
+      assert %IPDPacket{ciphertext: ciphertext} =
+               Encrypt.encrypt(%IPDPacket{plaintext: "Hello!"},
+                 session_key: sym_key,
+                 session_key_algo: @algo,
+                 use_mdc: true
+               )
 
       assert %IPDPacket{plaintext: "Hello!"} =
                IPDPacket.decrypt(
@@ -36,7 +73,13 @@ defmodule OpenPGP.IntegrityProtectedDataPacketTest do
     @algo {9, "AES with 256-bit key"}
     test "encrypt plaintext with AES-256" do
       sym_key = :crypto.strong_rand_bytes(32)
-      assert ciphertext = IPDPacket.encrypt("Hello!", sym_key, @algo, use_mdc: true)
+
+      assert %IPDPacket{ciphertext: ciphertext} =
+               Encrypt.encrypt(%IPDPacket{plaintext: "Hello!"},
+                 session_key: sym_key,
+                 session_key_algo: @algo,
+                 use_mdc: true
+               )
 
       assert %IPDPacket{plaintext: "Hello!"} =
                IPDPacket.decrypt(
@@ -50,7 +93,12 @@ defmodule OpenPGP.IntegrityProtectedDataPacketTest do
     test "encrypt plaintext with AES-128 and no MDC" do
       sym_key = :crypto.strong_rand_bytes(16)
 
-      assert ciphertext = IPDPacket.encrypt("Hello!", sym_key, @algo, use_mdc: false)
+      assert %IPDPacket{ciphertext: ciphertext} =
+               Encrypt.encrypt(%IPDPacket{plaintext: "Hello!"},
+                 session_key: sym_key,
+                 session_key_algo: @algo,
+                 use_mdc: false
+               )
 
       assert %IPDPacket{plaintext: "Hello!"} =
                IPDPacket.decrypt(
@@ -63,7 +111,13 @@ defmodule OpenPGP.IntegrityProtectedDataPacketTest do
     @algo {7, "AES with 128-bit key [AES]"}
     test "encrypt plaintext with AES-128 and no MDC (default behavior of .decrypt/2)" do
       sym_key = :crypto.strong_rand_bytes(16)
-      assert ciphertext = IPDPacket.encrypt("Hello!", sym_key, @algo, use_mdc: false)
+
+      assert %IPDPacket{ciphertext: ciphertext} =
+               Encrypt.encrypt(%IPDPacket{plaintext: "Hello!"},
+                 session_key: sym_key,
+                 session_key_algo: @algo,
+                 use_mdc: false
+               )
 
       assert %IPDPacket{plaintext: "Hello!"} =
                IPDPacket.decrypt(